% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvLSW.R
\name{AdjPositiveDef}
\alias{AdjPositiveDef}
\title{Adjust Matrix Estimate to be Strictly Positive Definite}
\usage{
AdjPositiveDef(object, tol = 1e-10)
}
\arguments{
\item{object}{An mvLSW object containing the multivariate evolutionary wavelet spectrum estimate.}

\item{tol}{The tolerance. Default value is \code{tol=1e-10}.}
}
\value{
An updated mvLSW object.
}
\description{
Adjust Matrix Estimate to be Strictly Positive Definite.
}
