% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mvLSW.R
\docType{methods}
\name{plot.mvLSW}
\alias{plot.mvLSW}

\title{Plot mvLSW Object}
\description{
  Plot the data contained within a \code{mvLSW} object based on the 
  requested format.
}
\usage{
  \method{plot}{mvLSW}(x, style = 1, info = NULL, Interval = NULL, 
    diag = TRUE, sub = "Spectrum", ...)
}

\arguments{
  \item{x}{A \code{mvLSW} object.}
  \item{style}{Index stating the type of plotting format for 
    the \code{mvLSW} object. (See details.)}
  \item{info}{Vector containing the channel and/or level indices 
    defining the slice through \code{x} according to the 
	requested plotting \code{style}. (See details.)}
  \item{Interval}{A list containing two items, both \code{mvLSW} 
    objects with names "L" and "U" that respectively define the lower
    and upper pointwise interval values. If \code{NULL}, default, 
	then no interval is plotted.}
  \item{diag}{Logical, should the diagonal panels be drawn when 
    \code{style=2}. Ideally this should be \code{FALSE} if 
	\code{object} contains the coherence. Set to \code{TRUE} by default.}
  \item{sub}{Plot subtitle. Set to \code{"Spectrum"} by default.}
  \item{...}{Additional graphical parameters.}
}

\details{
  This command plots the data contained within the \code{mvLSW} based 
  on requested plotting style.
  
  Plotting style \code{style=1} with information \code{info=c(p,q,j)} 
  generates a single plot for a 
  specified channel pair \code{p} & \code{q} and level \code{j}.
  
  Plotting style \code{style=2} with information \code{info=j} 
  creates a set of plots from \code{x} for all channel pairs in a 
  lower-triangular panel corresponding to the specified level j.
  If \code{diag=FALSE} then the plots along the diagonal are suppressed,
  which is ideal when \code{x} contain coherence estimates.

  Plotting style \code{style=3} with information \code{info=c(p,q)}
  creates a set of plots from \code{x} for all levels (from fine 
  to coarse) for channel pair \code{p} and \code{q}.
  
  Finally, the plotting style \code{style=4} with information 
  \code{info=c(p,q)} presents the same information as
  for the previous case, but in a compact matrix format. Please 
  refer to \code{image.plot} from the \code{fields} library for 
  additional information on this plotting style.
  
  The argument \code{Interval} must be supplied in order to draw a 
  polygon depicting the pointwise interval. See \code{\link{ApxCI}} 
  for deriving an approximate confidence interval for the evolutionary
  wavelet spectrum estimate.
  This argument is ignored in the case \code{style=4}.
}

\value{
  Generates a plot. No data is returned.
}

\references{
  Taylor, S.A.C., Park, T.A. and Eckley, I. (2019) Multivariate 
  locally stationary wavelet analysis with the mvLSW R package.
  \emph{Journal of statistical software} \strong{90}(11) pp. 1--16,
  doi: 10.18637/jss.v090.i11.
}

\seealso{
  \code{plot.default}, \code{image.plot}, \code{\link{as.mvLSW}}, 
  \code{\link{mvEWS}}, \code{\link{coherence}}, \code{\link{ApxCI}}.
}

\examples{
## Define evolutionary wavelet spectrum, structure only on level 2
Spec <- array(0, dim=c(3, 3, 8, 256))
Spec[1, 1, 2, ] <- 10
Spec[2, 2, 2, ] <- c(rep(5, 64), rep(0.6, 64), rep(5, 128))
Spec[3, 3, 2, ] <- c(rep(2, 128), rep(8, 128))
Spec[2, 1, 2, ] <- Spec[1, 2, 2, ] <- punif(1:256, 65, 192)
Spec[3, 1, 2, ] <- Spec[1, 3, 2, ] <- c(rep(-1, 128), rep(5, 128))
Spec[3, 2, 2, ] <- Spec[2, 3, 2, ] <- -0.5
EWS <- as.mvLSW(x = Spec, filter.number = 1, family = "DaubExPhase",
  min.eig.val = NA)

## Sample time series and estimate the EWS and coherence.
set.seed(10)
X <- rmvLSW(Spectrum = EWS)
EWS_X <- mvEWS(X, kernel.name = "daniell", kernel.param = 20)
RHO_X <- coherence(EWS_X, partial = FALSE)

## Evaluate asymptotic spectral variance 
SpecVar <- varEWS(EWS_X)

## Evaluate 95% approximate confidence interval
CI <- ApxCI(object = EWS_X, var = SpecVar, alpha=0.05)

## Plot mvEWS between channels 1 & 3 at level 2
plot(x = EWS_X, style = 1, info = c(1, 3, 2), Interval = CI)

## Plot coherence between channels 1 & 3 at level 2
plot(x = RHO_X, style = 1, info = c(1, 3, 2), ylab = "Coherence")

## mvEWS panel plot for level 2
plot(x = EWS_X, style = 2, info = 2, Interval = CI)

## Panel plot of coherence for level 2
plot(x = RHO_X, style = 2, info = 2, diag = FALSE, ylab = "Coherence")

## Plot mvEWS for channel pair 1 & 3 at all levels
plot(x = EWS_X, style = 3, info = c(1, 3), Interval = CI)

## Plot coherence for channel pair 1 & 3 at all levels
plot(x = RHO_X, style = 3, info = c(1, 3), ylab = "Coherence")

## Image plot for coherence between channels 1 & 3
plot(x = RHO_X, style = 4, info = c(1, 3), sub = "Coherence")
}

\keyword{plot.mvLSW}
