% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvLSW.R
\name{TODO *************}
\alias{TODO *************}
\alias{Smooth_EWS}
\title{Smoothining of the Evolutionary Wavelet Spectrum Estimate (wrapper)}
\usage{
Smooth_EWS(
  object,
  kernel.name = "daniell",
  kernel.param = 1,
  optimize = FALSE,
  type = "all",
  level = NA
)
}
\arguments{
\item{object}{An mvLSW object containing the multivariate evolutionary wavelet spectrum estimate.}

\item{kernel.name}{Name of smoothing kernel to be supplied to \code{kernel()}. Kernel \code{"daniell"} is defined by default.}

\item{kernel.param}{Parameters to be passed to \code{kernel()}. This argument must be a vector if \code{type="all"}, otherwise it must be a matrix with each column defining the kernel parameters for each \code{log2(nrow(X))} levels from coarse to fine. If the name is \code{"dirichlet"} or \code{"fejer"} then \code{kernel.param} must have length 2 (or a matrix with 2 rows) which are supplied to \code{kernel()} as arguments \code{m} and \code{r} respectively. Note that the width of the kernel cannot be larger than the time series length. This is set by default as the square root of the length of the time series.}

\item{optimize}{Logical, should the smoothing be optimized? If \code{FALSE} (default) then smoothing is performed as specified with \code{kernel.name} and \code{kernel.param}.  Otherwise, \code{kernel.param} defines the upper parameter bound in determining the optimal kernel is determined by minimising the generalized cross-validation gamma deviance criterion.
@param type How should the smoothing be performed? If \code{"all"} (default) then the same smoothing kernel is applied to all levels, else if \code{"by.level"} then a different smoothing#'  kernel is applied to each level.}

\item{type}{TODO !!!!}

\item{level}{TODO !!!!}
}
\value{
An object of class \code{mvLSW}, invisibly.
}
\description{
Smoothining of the Evolutionary Wavelet Spectrum Estimate
}
